+++ wrappers/python/eduvpn_common/loader.py
@@ -112,9 +112,6 @@ def initialize_functions(lib: CDLL) -> None:
         c_char_p,
         c_char_p,
     ], c_void_p
-    lib.SetSupportWireguard.argtypes, lib.SetSupportWireguard.restype = [
-        c_int,
-    ], c_void_p
     lib.SetState.argtypes, lib.SetState.restype = [
         c_int,
     ], c_void_p