+++ wrappers/python/eduvpn_common/main.py
@@ -328,18 +328,6 @@ class EduVPN(object):
         if renew_err:
             forwardError(renew_err)
 
-    def set_support_wireguard(self, support: bool) -> None:
-        """Indicates whether or not the OS supports WireGuard connections.
-
-        :param support: bool: whether or not wireguard is supported
-
-        :raises WrappedError: An error by the Go library
-        """
-        support_err = self.go_function(self.lib.SetSupportWireguard, support)
-
-        if support_err:
-            forwardError(support_err)
-
     def start_failover(
         self, gateway: str, wg_mtu: int, readrxbytes: ReadRxBytes
     ) -> bool: